First-generation topological selection autodetect (#1575)
* Separate V0 and cascades

* Add extra files

* First-generation selection autodetect added

* Please consider the following formatting changes
