+++ PWGLF/TableProducer/CMakeLists.txt
@@ -32,10 +32,16 @@ o2physics_add_dpl_workflow(cascadebuilder
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore O2::DetectorsVertexing
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(strangenessbuilder
-                    SOURCES strangenessbuilder.cxx
+o2physics_add_dpl_workflow(singlestrangebuilder
+                    SOURCES singlestrangebuilder.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore O2::DetectorsVertexing
                     COMPONENT_NAME Analysis)
+
+o2physics_add_dpl_workflow(multistrangebuilder
+                    SOURCES multistrangebuilder.cxx
+                    PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsBase O2Physics::AnalysisCore O2::DetectorsVertexing
+                    COMPONENT_NAME Analysis)
+
 
 o2physics_add_dpl_workflow(cascadefinder
                     SOURCES cascadefinder.cxx
